Satisfy checkstyle conditions mentioned in se-education.org
